+++ src/index.tsx
@@ -11,8 +11,8 @@ import { ReactQueryDevtools } from 'react-query/devtools';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router } from 'react-router-dom';
 
+import { SubstrateProvider } from '@/lib/substrate';
 import SubstrateLoadingAndErrorHandlingWrapper from '@/lib/substrate/components/SubstrateLoadingAndErrorHandlingWrapper';
-import { SubstrateProvider } from '@/lib/substrate/substrate-context';
 
 import App from './App';
 import reportWebVitals from './reportWebVitals';
