+++ src/lib/substrate/index.ts
@@ -1,8 +1,8 @@
-import { SubstrateProvider, useSubstrate, useSubstrateSecureState, useSubstrateState } from './substrate-context';
 import utils from './utils';
 
 // ray test touch <
 export * from './context/types';
-// ray test touch >
+export * from './substrate-context';
 
-export { SubstrateProvider, utils as substrateUtils, useSubstrate, useSubstrateSecureState, useSubstrateState };
+export { utils as substrateUtils };
+// ray test touch >
